project/requirements.txt
pygame
neat-python
numpy
pillow
graphviz
matplotlib